+++ OpenHD/ohd_video/src/air_recording_helper.hpp
