+++ OpenHD/ohd_video/src/gstreamerstream.cpp
@@ -7,7 +7,7 @@
 #include <regex>
 #include <vector>
 
-#include "AirRecordingFileHelper.hpp"
+#include "air_recording_helper.hpp"
 #include "ffmpeg_videosamples.hpp"
 #include "gst_helper.hpp"
